calcularSubtotal convierte parámetro a Int

+++ app/src/main/java/cl/dbustamante/android/progiitarea/models/ItemMesa.kt
@@ -5,6 +5,6 @@ class ItemMesa(
     private val cantidad: Int
 ) {
     fun calcularSubtotal(): Int {
-        return itemMenu.precio * cantidad
+        return itemMenu.precio.toInt() * cantidad
     }
 }
